+++ NodeExpr.java
@@ -1,3 +1,7 @@
+/**
+ * Represents an expression node, which is a sequence of terms
+ * separated by addops (e.g., "a + b - c").
+ */
 public class NodeExpr extends Node {
 
 	private NodeTerm term;
@@ -10,6 +14,11 @@ public class NodeExpr extends Node {
 		this.expr=expr;
 	}
 
+	/**
+	 * Appends a new term/addop to the expression, maintaining
+	 * left-associativity for evaluation.
+	 * @param expr The expression part to append.
+	 */
 	public void append(NodeExpr expr) {
 		if (this.expr==null) {
 			this.addop=expr.addop;
@@ -19,12 +28,22 @@ public class NodeExpr extends Node {
 			this.expr.append(expr);
 	}
 
-	public int eval(Environment env) throws EvalException {
+	/**
+	 * Evaluates the expression (handles left-associativity).
+	 * @param env The environment to evaluate in.
+	 * @return The double result of the expression.
+	 * @throws EvalException If evaluation fails.
+	 */
+	public double eval(Environment env) throws EvalException {
 		return expr==null
 			? term.eval(env)
 			: addop.op(expr.eval(env),term.eval(env));
 	}
 
+	/**
+	 * Generates C code for the expression.
+	 * @return A string of C code.
+	 */
 	public String code() {
 		return (expr==null ? "" : expr.code()+addop.code())+term.code();
 	}
