+++ NodeStmt.java
@@ -1,3 +1,6 @@
+/**
+ * Represents a statement, which in TA1 is just an assignment.
+ */
 public class NodeStmt extends Node {
 
 	private NodeAssn assn;
@@ -6,10 +9,20 @@ public class NodeStmt extends Node {
 		this.assn=assn;
 	}
 
-	public int eval(Environment env) throws EvalException {
+	/**
+	 * Evaluates the underlying assignment.
+	 * @param env The environment to evaluate in.
+	 * @return The result of the assignment expression.
+	 * @throws EvalException If the assignment fails.
+	 */
+	public double eval(Environment env) throws EvalException {
 		return assn.eval(env);
 	}
 
+	/**
+	 * Generates C code for the statement.
+	 * @return C code for the assignment.
+	 */
 	public String code() { return assn.code(); }
 
 }
